refactor: move env-and-consts.ts to the common folder rather than in the renderer folder (#120)
