+++ src/renderer/components/configure-recording-set-page/set-meta-configuration.tsx
@@ -3,7 +3,7 @@ import Col from 'react-bootstrap/Col';
 import Form from 'react-bootstrap/Form';
 import { useTranslation } from 'react-i18next';
 
-import { SCALE_KEYS_IN_ORDER, SUPPORTED_OCTAVES_IN_ORDER } from '../../env-and-consts';
+import { SCALE_KEYS_IN_ORDER, SUPPORTED_OCTAVES_IN_ORDER } from '../../../common/env-and-consts';
 import { Consumer, RecordingSet, ScaleKey, SupportedOctave } from '../../types';
 import { Select } from '../select';
