+++ src/renderer/contexts/recording-project-context.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { noOp } from '../env-and-consts';
+import { noOp } from '../../common/env-and-consts';
 import { Consumer, RecordingProject } from '../types';
 
 interface RecordingProjectContextType {
