+++ src/renderer/services/i18n/index.ts
@@ -2,7 +2,7 @@ import i18n from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import { initReactI18next } from 'react-i18next';
 
-import { KEY_SEPARATOR, LOCAL_STORAGE_KEY_LOCALE, isDevelopment } from '../../env-and-consts';
+import { KEY_SEPARATOR, LOCAL_STORAGE_KEY_LOCALE, isDevelopment } from '../../../common/env-and-consts';
 
 import WebpackImportBackend from './webpack-import-backend';
